check_bot/requirements.txt
aerich==0.7.2
aiosqlite==0.17.0
annotated-types==0.5.0
anyio==3.7.1
asyncpg==0.28.0
beautifulsoup4==4.12.3
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cycler==0.11.0
dictdiffer==0.9.0
exceptiongroup==1.2.0
fonttools==4.38.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.6
importlib-metadata==6.7.0
iso8601==1.1.0
kiwisolver==1.4.5
matplotlib==3.5.3
numpy==1.21.6
opencv-python==4.9.0.80
packaging==23.2
pandas==1.3.5
Pillow==9.5.0
psutil==5.9.8
py-cpuinfo==9.0.0
pydantic==2.5.3
pydantic_core==2.14.6
pyparsing==3.1.1
pypika-tortoise==0.1.6
python-dateutil==2.8.2
python-telegram-bot==20.3
pytz==2023.3.post1
PyYAML==6.0.1
pyzbar==0.1.9
qrdet==2.4
qreader==3.12
quadrilateral-fitter==1.8
requests==2.31.0
scipy==1.7.3
seaborn==0.12.2
shapely==2.0.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
tomlkit==0.12.3
torch==1.13.1
torchvision==0.14.1
tortoise-orm==0.19.3
tqdm==4.66.1
typing_extensions==4.7.1
ultralytics==8.0.145
urllib3==2.0.7
zipp==3.15.0